requirements.txt
# Core Library Dependencies (paper2slides/)
# RAG Core
huggingface_hub
lightrag-hku
mineru[core]==2.6.4
tqdm

# Image Processing
Pillow>=10.0.0
python-pptx>=1.0.0
reportlab>=4.0.0

# API
openai>=1.0.0
python-dotenv>=1.0.0

# Data Processing
pyyaml>=6.0
requests>=2.28.0
markdown>=3.7


# Web API Dependencies (api/)
fastapi==0.122
uvicorn[standard]==0.38.0
python-multipart==0.0.20
pydantic==2.12.3

httpx>=0.27.0

# Optional: PGF rendering (requires poppler system package)
# pdf2image>=2.7.0
